refactor of razor files
