+++ src/Chirp.Web/Pages/Public.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@model Chirp.Web.Razor.PublicModel
+@model Chirp.Web.Razor.PublicTimeLine
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "./Shared/_Layout.cshtml";
