+++ src/Chirp.Web/Pages/Shared/_Layout.cshtml
@@ -3,6 +3,7 @@
 <head>
     <title>@ViewData["Title"]</title>
     <link href="/css/style.css" rel="stylesheet" type="text/css" />
+    <link href="/css/index.css" rel=stylesheet" type="text/css"/>
     <link rel="icon" type="image/png" href="/favicon/favicon.ico">
 </head>
 <body>
