+++ src/Chirp.Web/Razor/PublicTimeline.cshtml.cs
@@ -7,36 +7,8 @@ using Chirp.Core.DTO;
 using System.Threading.Tasks;
 #pragma warning disable CS8604 // Possible null reference argument.
 
-public class PublicModel(ICheepService cheepService) : PageModel
+public class PublicTimeLine(ICheepService cheepService) : TimeLine(cheepService)
 {
-    private readonly ICheepService _cheepService = cheepService;
-
-    public required List<CheepDTO> Cheeps { get; set; }
-    public required List<AuthorDTO> Authors { get; set; }
-    public int Count {get; set; }
-    public int NextPage {get; set;}
-    public int PreviousPage {get; set;}
-    public int CurrentPage {get; set;}
-
-    public int LastPage {get; set;}
-
-
-    [BindProperty(SupportsGet = true)]
-     public string? SearchName {get; set;}
-
-    public int DefinePreviousPage(int page){
-        if(page == 0){
-            return 0;
-        } else{
-            return page-1;
-        }
-    }
-
-    public int DefineLastPage(){
-        double p = Count/32;
-        int LastPage = (int) Math.Ceiling(p);
-        return LastPage;
-    }
 
     public async Task<ActionResult> OnGetAsync(int page = 0)
     {
@@ -56,5 +28,6 @@ public class PublicModel(ICheepService cheepService) : PageModel
         LastPage = DefineLastPage();
         return Page();
     }
+    
     
 }
